[GridNG] Returning auto repeat named line index for subgrid
This CL adds an extra condition to only check for the auto repeat track
list length when placing items in a standalone grid. Because in a valid
definition of a subgrid's named lines there shouldn't be any tracks,
|FirstExplicitPosition| in |NamedLineCollection| would try to return an
index of a named line outside of a repeat and crash when such lines
weren't defined. A web test was added to validate this behavior.

Note that further examination of this behavior will be needed once it's
determined how the line names will be collected for subgrid.

Bug: 1294472
Change-Id: Idde6662108cda10ceea5218c0440ca36dd51490b